requirements.txt
PyYAML==6.0
loguru==0.7.2
